\name{statistics.general-package}
\alias{statistics.general-package}
\alias{statistics.general}
\docType{package}
\title{
Statistics General
}
\description{
A common set of general statistics functions
}
\details{
\tabular{ll}{
Package: \tab statistics.general\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-04\cr
License: \tab GPL-2\cr
}
}
\author{
Richard de Borja <richard.deborja@sickkids.ca>
}
\keyword{ statistics }
